refactor(DashAdmin): se reestructuro y arreglaron archivos
