+++ frontend/app/dashboard/Admin/assets/css/CreateUserStyle.css
@@ -1,0 +1,338 @@
+/* Estilos específicos para el formulario de creación de usuarios */
+
+/* Contenedor principal */
+.create-user-container {
+    width: 100%;
+    max-width: 100%;
+    padding: 1rem;
+}
+
+/* Breadcrumb */
+.breadcrumb {
+    margin-bottom: 2rem;
+}
+
+.breadcrumb-link {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.5rem;
+    color: #d4d4d4;
+    text-decoration: none;
+    font-size: 0.9rem;
+    transition: color 0.2s ease;
+}
+
+.breadcrumb-link:hover {
+    color: #ffffff;
+}
+
+.breadcrumb-link img {
+    width: 1rem;
+    height: 1rem;
+    filter: brightness(0) invert(1);
+    opacity: 0.7;
+}
+
+/* Contenedor del formulario */
+.form-container {
+    background: rgba(0, 0, 0, 0.15);
+    backdrop-filter: blur(15px);
+    border-radius: 1rem;
+    padding: 2rem;
+    border: 1px solid rgba(0, 0, 0, 0.623);
+    box-shadow: 
+        0 30px 60px rgba(0, 0, 0, 0.3),
+        inset 0 1px 0 rgba(131, 131, 131, 0.2);
+}
+
+/* Header del formulario */
+.form-header {
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+    margin-bottom: 2rem;
+    padding-bottom: 1rem;
+    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.form-icon {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 3rem;
+    height: 3rem;
+    background-color: rgba(255, 255, 255, 0.1);
+    border-radius: 0.8rem;
+    flex-shrink: 0;
+}
+
+.form-icon img {
+    width: 1.5rem;
+    height: 1.5rem;
+    filter: brightness(0) invert(1);
+}
+
+.form-title h2 {
+    font-size: 1.5rem;
+    font-weight: 600;
+    color: #f3f3f3;
+    margin-bottom: 0.3rem;
+}
+
+.form-title p {
+    color: #d4d4d4;
+    font-size: 0.9rem;
+}
+
+/* Grid del formulario */
+.form-grid {
+    display: grid;
+    gap: 2rem;
+}
+
+/* Secciones del formulario */
+.form-section {
+    background: rgba(0, 0, 0, 0.2);
+    border-radius: 0.8rem;
+    padding: 1.5rem;
+    border: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.section-title {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    font-size: 1.1rem;
+    font-weight: 600;
+    color: #f3f3f3;
+    margin-bottom: 1.5rem;
+}
+
+.section-title img {
+    width: 1.2rem;
+    height: 1.2rem;
+    filter: brightness(0) invert(1);
+    opacity: 0.8;
+}
+
+/* Filas del formulario */
+.form-row {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    gap: 1rem;
+    margin-bottom: 1rem;
+}
+
+.form-row:last-child {
+    margin-bottom: 0;
+}
+
+/* Grupos de campos */
+.form-group {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.form-group.full-width {
+    grid-column: 1 / -1;
+}
+
+.form-group label {
+    font-size: 0.9rem;
+    font-weight: 500;
+    color: #f3f3f3;
+}
+
+.form-group input,
+.form-group select {
+    padding: 0.8rem 1rem;
+    border: 1px solid rgba(255, 255, 255, 0.2);
+    border-radius: 0.5rem;
+    background-color: rgba(255, 255, 255, 0.1);
+    color: #fff;
+    font-size: 0.9rem;
+    transition: all 0.2s ease;
+}
+
+.form-group input::placeholder {
+    color: rgba(255, 255, 255, 0.6);
+}
+
+.form-group input:focus,
+.form-group select:focus {
+    outline: none;
+    border-color: rgba(255, 255, 255, 0.4);
+    background-color: rgba(255, 255, 255, 0.15);
+    box-shadow: 0 0 0 3px rgba(255, 255, 255, 0.1);
+}
+
+.form-group select {
+    cursor: pointer;
+}
+
+.form-group select option {
+    background-color: #2a2a2a;
+    color: #fff;
+}
+
+/* Estados de error */
+.form-group input.error,
+.form-group select.error {
+    border-color: rgba(231, 76, 60, 0.6);
+    background-color: rgba(231, 76, 60, 0.1);
+}
+
+.form-group input.error:focus,
+.form-group select.error:focus {
+    border-color: rgba(231, 76, 60, 0.8);
+    box-shadow: 0 0 0 3px rgba(231, 76, 60, 0.2);
+}
+
+/* Botones de acción */
+.form-actions {
+    display: flex;
+    justify-content: flex-end;
+    gap: 1rem;
+    margin-top: 2rem;
+    padding-top: 1.5rem;
+    border-top: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.btn-cancel,
+.btn-submit {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.5rem;
+    padding: 0.8rem 1.5rem;
+    border-radius: 0.6rem;
+    text-decoration: none;
+    font-weight: 600;
+    font-size: 0.9rem;
+    transition: all 0.2s ease;
+    border: none;
+    cursor: pointer;
+}
+
+.btn-cancel {
+    background-color: rgba(108, 117, 125, 0.2);
+    color: #d4d4d4;
+    border: 1px solid rgba(108, 117, 125, 0.3);
+}
+
+.btn-cancel:hover {
+    background-color: rgba(108, 117, 125, 0.3);
+    color: #ffffff;
+    transform: translateY(-1px);
+}
+
+.btn-submit {
+    background-color: #ffffff;
+    color: #000;
+    border: 1px solid transparent;
+}
+
+.btn-submit:hover {
+    background-color: #f0f0f0;
+    transform: translateY(-1px);
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+}
+
+.btn-cancel img,
+.btn-submit img {
+    width: 1rem;
+    height: 1rem;
+}
+
+.btn-cancel img {
+    filter: brightness(0) invert(1);
+}
+
+.btn-submit img {
+    filter: brightness(0);
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+    .create-user-container {
+        padding: 0.5rem;
+    }
+    
+    .form-container {
+        padding: 1.5rem;
+    }
+    
+    .form-header {
+        flex-direction: column;
+        text-align: center;
+        gap: 1rem;
+    }
+    
+    .form-row {
+        grid-template-columns: 1fr;
+    }
+    
+    .form-actions {
+        flex-direction: column-reverse;
+        gap: 0.8rem;
+    }
+    
+    .btn-cancel,
+    .btn-submit {
+        width: 100%;
+        justify-content: center;
+    }
+}
+
+@media (max-width: 480px) {
+    .form-container {
+        padding: 1rem;
+    }
+    
+    .form-section {
+        padding: 1rem;
+    }
+    
+    .section-title {
+        font-size: 1rem;
+    }
+    
+    .form-group input,
+    .form-group select {
+        padding: 0.7rem 0.8rem;
+        font-size: 0.8rem;
+    }
+}
+
+/* Animaciones */
+@keyframes fadeIn {
+    from {
+        opacity: 0;
+        transform: translateY(10px);
+    }
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.form-container {
+    animation: fadeIn 0.3s ease;
+}
+
+.form-section {
+    animation: fadeIn 0.4s ease;
+}
+
+/* Mejoras de accesibilidad */
+.form-group input:focus-visible,
+.form-group select:focus-visible {
+    outline: 2px solid rgba(255, 255, 255, 0.5);
+    outline-offset: 2px;
+}
+
+.btn-cancel:focus-visible,
+.btn-submit:focus-visible {
+    outline: 2px solid rgba(255, 255, 255, 0.5);
+    outline-offset: 2px;
+}
