+++ frontend/app/dashboard/Admin/assets/css/sidebar-right.css
@@ -1,0 +1,386 @@
+.sidebar-right {
+    width: 15rem;
+    height: 52.3rem;
+    position: fixed;
+    top: 0;
+    right: 0;
+    padding: 0.2rem;
+    border-radius: 1rem;
+    background-color: #222222;
+    box-shadow: -2px 0px 5px rgba(0, 0, 0, 0.1);
+    margin-top: 4.4rem; 
+}
+
+/* Contenedor de búsqueda */
+.search-container {
+    padding: 0.5rem ;
+    position: relative;
+    margin-bottom: 0.5rem;
+}
+
+/* Input de búsqueda */
+#searchInput {
+    width: 100%;
+    padding: 0.6rem 1rem 0.6rem 2rem;
+    border-radius: 0.5rem;
+    border: 1px solid #ccc;
+    font-size: 0.9rem;
+    background-color: #383838;
+    color: white;
+}
+
+#searchInput::placeholder {
+    color: #ccc;
+}
+
+/* Icono de búsqueda */
+.search-icon {
+    position: absolute;
+    left: 1.5rem;
+    top: 50%;
+    transform: translateY(-50%);
+    width: 1rem;
+    opacity: 0.6;
+}
+
+/* Contenedor de lotes */
+.lotes-container {
+    max-height: calc(55rem - 8rem);
+    overflow-y: auto;
+}
+
+/* Estilo de cada lote */
+.lote-item {
+    display: flex;
+    align-items: center;
+    gap: 0.8rem;
+    padding: 0.6rem 1rem;
+    text-decoration: none;
+    color: #e0e0e0;
+    transition: background-color 0.2s ease;
+}
+
+.lote-item:hover {
+    background-color: #333333;
+    cursor: pointer;
+}
+
+.lote-item img {
+    width: 1.2rem;
+    height: 1.2rem;
+    opacity: 0.8;
+}
+
+
+main {
+    margin-right: 16rem;
+    transition: margin-right 0.5s ease;
+}
+
+/* scrollbar */
+
+.sidebar-right ::-webkit-scrollbar {
+    background-color: transparent;
+    width: 10px;
+}
+
+.sidebar-right ::-webkit-scrollbar-thumb {
+    background-color: white;
+    border-radius: 3rem;
+}
+
+
+@media (max-width: 768px) {
+    .sidebar-right {
+        width: 100%;
+        transform: translateX(100%);
+        transition: transform 0.3s ease;
+    }
+    
+    .sidebar-right.active {
+        transform: translateX(0);
+    }
+    
+    main {
+        margin-right: 0;
+    }
+}
+
+/*lote-details.css*/
+
+/* Estilos para los elementos de lote en el sidebar */
+.lote-item {
+    display: flex;
+    align-items: center;
+    gap: 0.8rem;
+    padding: 0.8rem 1rem;
+    border-radius: 0.5rem;
+    transition: background-color 0.2s ease;
+    margin-bottom: 0.3rem;
+}
+
+.lote-item:hover {
+    background-color: #333333;
+}
+
+.lote-item.active {
+    background-color: #444444;
+    border-left: 3px solid #ffffff;
+}
+
+.lote-info {
+    display: flex;
+    flex-direction: column;
+}
+
+.lote-numero {
+    font-size: 0.9rem;
+    font-weight: 500;
+    color: #ffffff;
+}
+
+.lote-propietario {
+    font-size: 0.75rem;
+    color: #aaaaaa;
+}
+
+.status-ok {
+    color: #2ecc71;
+}
+
+.status-mora {
+    color: #e74c3c;
+}
+
+.status-vacant {
+    color: #95a5a6;
+}
+
+/* Estilos para el detalle del lote */
+.lote-detail-container {
+    padding: 1.5rem;
+    display: none;
+}
+
+.lote-detail-header {
+    display: flex;
+    align-items: center;
+    margin-bottom: 2rem;
+    flex-wrap: wrap;
+    gap: 1rem;
+}
+
+.back-button {
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+    padding: 0.5rem 1rem;
+    background-color: #f5f5f5;
+    border-radius: 2rem;
+    transition: background-color 0.2s ease;
+    margin-right: 1rem;
+}
+
+.back-button:hover {
+    background-color: #e0e0e0;
+}
+
+.back-button img {
+    width: 1rem;
+    margin-right: 0.5rem;
+}
+
+.lote-detail-header h1 {
+    font-size: 1.8rem;
+    font-weight: 600;
+    margin: 0;
+    flex-grow: 1;
+}
+
+.lote-status {
+    padding: 0.5rem 1rem;
+    border-radius: 2rem;
+    font-size: 0.9rem;
+    font-weight: 500;
+}
+
+.lote-status.al-dia {
+    background-color: rgba(46, 204, 113, 0.2);
+    color: #27ae60;
+}
+
+.lote-status.mora {
+    background-color: rgba(231, 76, 60, 0.2);
+    color: #e74c3c;
+}
+
+.lote-status.vacio {
+    background-color: rgba(149, 165, 166, 0.2);
+    color: #7f8c8d;
+}
+
+/* Grid para la información del lote */
+.lote-info-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 1.5rem;
+}
+
+.lote-info-card {
+    background-color: white;
+    border-radius: 0.8rem;
+    padding: 1.5rem;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);
+}
+
+.lote-info-card h2 {
+    font-size: 1.2rem;
+    font-weight: 600;
+    margin-bottom: 1.5rem;
+    color: #333;
+    border-bottom: 1px solid #f0f0f0;
+    padding-bottom: 0.8rem;
+}
+
+.info-group {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 1rem;
+    align-items: center;
+}
+
+.info-label {
+    font-size: 0.9rem;
+    color: #777;
+    font-weight: 500;
+}
+
+.info-value {
+    font-size: 0.95rem;
+    color: #333;
+    font-weight: 500;
+    text-align: right;
+}
+
+.info-value.positive {
+    color: #27ae60;
+}
+
+.info-value.negative {
+    color: #e74c3c;
+}
+
+/* Tabla de historial de pagos */
+.payments-table {
+    width: 100%;
+    border-collapse: collapse;
+}
+
+.payments-table th {
+    padding: 0.8rem;
+    text-align: left;
+    font-size: 0.85rem;
+    color: #777;
+    font-weight: 500;
+    border-bottom: 1px solid #f0f0f0;
+}
+
+.payments-table td {
+    padding: 0.8rem;
+    font-size: 0.85rem;
+    border-bottom: 1px solid #f0f0f0;
+}
+
+.estado-pagado {
+    color: #27ae60;
+    background-color: rgba(46, 204, 113, 0.1);
+    padding: 0.3rem 0.6rem;
+    border-radius: 1rem;
+    font-size: 0.75rem;
+}
+
+.estado-pendiente {
+    color: #e74c3c;
+    background-color: rgba(231, 76, 60, 0.1);
+    padding: 0.3rem 0.6rem;
+    border-radius: 1rem;
+    font-size: 0.75rem;
+}
+
+/* Botones de acción rápida */
+.quick-actions {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 1rem;
+}
+
+.action-button {
+    padding: 0.8rem;
+    border: none;
+    border-radius: 0.5rem;
+    font-size: 0.9rem;
+    font-weight: 500;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    color: white;
+}
+
+.action-button.registrar-pago {
+    background-color: #3a7bd5;
+}
+
+.action-button.registrar-pago:hover {
+    background-color: #2c5c9c;
+}
+
+.action-button.generar-expensa {
+    background-color: #27ae60;
+}
+
+.action-button.generar-expensa:hover {
+    background-color: #219653;
+}
+
+.action-button.enviar-notificacion {
+    background-color: #f39c12;
+}
+
+.action-button.enviar-notificacion:hover {
+    background-color: #d68910;
+}
+
+.action-button.ver-documentos {
+    background-color: #9b59b6;
+}
+
+.action-button.ver-documentos:hover {
+    background-color: #8e44ad;
+}
+
+/* Mensajes de no hay datos */
+.no-data {
+    text-align: center;
+    padding: 1.5rem;
+    color: #777;
+    font-style: italic;
+}
+
+/* Media Queries */
+@media (max-width: 768px) {
+    .lote-info-grid {
+        grid-template-columns: 1fr;
+    }
+    
+    .quick-actions {
+        grid-template-columns: 1fr;
+    }
+    
+    .lote-detail-header {
+        flex-direction: column;
+        align-items: flex-start;
+    }
+    
+    .back-button {
+        margin-bottom: 1rem;
+    }
+}
