--- frontend/app/dashboard/Admin/lotes/Users/UserlistStyle.css
@@ -1,388 +1,0 @@
-/* Estilos específicos para la lista de usuarios */
-
-/* Contenedor principal */
-.usuarios-container {
-    width: 100%;
-    max-width: 100%;
-    padding: 1rem;
-}
-
-/* Sección de acciones rápidas */
-.actions-container {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-    gap: 1.5rem;
-    margin-bottom: 2rem;
-}
-
-.action-card {
-    display: flex;
-    align-items: center;
-    background: rgba(0, 0, 0, 0.15);
-    backdrop-filter: blur(15px);
-    border-radius: 1rem;
-    padding: 1.2rem;
-    border: 1px solid rgba(0, 0, 0, 0.623);
-    box-shadow: 
-        0 30px 60px rgba(0, 0, 0, 0.3),
-        inset 0 1px 0 rgba(131, 131, 131, 0.2);
-    transition: all 0.3s ease;
-}
-
-.action-card:hover {
-    transform: translateY(-2px);
-    box-shadow: 0 8px 25px rgba(0, 0, 0, 0.15);
-}
-
-.action-icon {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 3rem;
-    height: 3rem;
-    border-radius: 0.8rem;
-    margin-right: 1rem;
-    background-color: black;
-    flex-shrink: 0;
-}
-
-.action-icon img {
-    width: 1.5rem;
-    height: 1.5rem;
-    filter: brightness(0) invert(1);
-}
-
-.action-info {
-    flex: 1;
-}
-
-.action-info h3 {
-    font-size: 1rem;
-    font-weight: 600;
-    margin-bottom: 0.3rem;
-    color: #f3f3f3;
-}
-
-.action-info p {
-    font-size: 0.8rem;
-    color: #d4d4d4;
-}
-
-.action-button {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 2.5rem;
-    height: 2.5rem;
-    background-color: #ffffff;
-    border-radius: 0.6rem;
-    text-decoration: none;
-    transition: all 0.2s ease;
-    flex-shrink: 0;
-}
-
-.action-button:hover {
-    background-color: #f0f0f0;
-    transform: scale(1.05);
-}
-
-.action-button img {
-    width: 1.2rem;
-    height: 1.2rem;
-    filter: brightness(0);
-}
-
-.newuser {
-    text-decoration: solid;
-}
-
-/* Input de búsqueda en la tarjeta */
-.search-input-container {
-    display: flex;
-    align-items: center;
-    flex-shrink: 0;
-}
-
-#searchUsers {
-    padding: 0.5rem 0.8rem;
-    border: 1px solid rgba(255, 255, 255, 0.2);
-    border-radius: 0.5rem;
-    background-color: rgba(255, 255, 255, 0.1);
-    color: #fff;
-    font-size: 0.9rem;
-    width: 150px;
-    transition: all 0.2s ease;
-}
-
-#searchUsers::placeholder {
-    color: rgba(255, 255, 255, 0.6);
-}
-
-#searchUsers:focus {
-    outline: none;
-    border-color: rgba(255, 255, 255, 0.4);
-    background-color: rgba(255, 255, 255, 0.15);
-}
-
-/* Contador de usuarios */
-.total-users {
-    font-size: 0.9rem;
-    color: #dfdfdf;
-    background: rgba(0, 0, 0, 0.2);
-    padding: 0.3rem 0.8rem;
-    border-radius: 1rem;
-    border: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-/* Contenedor de la tabla */
-.usuarios-table-container {
-    background: rgba(0, 0, 0, 0.15);
-    backdrop-filter: blur(15px);
-    border-radius: 1rem;
-    padding: 1.5rem;
-    border: 1px solid rgba(0, 0, 0, 0.623);
-    box-shadow: 
-        0 30px 60px rgba(0, 0, 0, 0.3),
-        inset 0 1px 0 rgba(131, 131, 131, 0.2);
-}
-
-.table-wrapper {
-    overflow-x: auto;
-}
-
-/* Estilos de la tabla */
-.usuarios-table {
-    width: 100%;
-    border-collapse: collapse;
-    background: transparent;
-}
-
-.usuarios-table thead {
-    background: rgba(0, 0, 0, 0.3);
-}
-
-.usuarios-table th {
-    padding: 1rem;
-    text-align: left;
-    border-bottom: 2px solid rgba(255, 255, 255, 0.1);
-    color: #f3f3f3;
-    font-weight: 600;
-    font-size: 0.9rem;
-}
-
-.th-content {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-}
-
-.th-content img {
-    width: 1rem;
-    height: 1rem;
-    filter: brightness(0) invert(1);
-    opacity: 0.7;
-}
-
-.usuarios-table td {
-    padding: 1rem;
-    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-    color: #e0e0e0;
-    font-size: 0.9rem;
-}
-
-.usuario-row {
-    transition: all 0.2s ease;
-}
-
-.usuario-row:hover {
-    background: rgba(255, 255, 255, 0.05);
-}
-
-.usuario-id {
-    font-weight: 600;
-    color: #ffffff;
-}
-
-.usuario-nombre,
-.usuario-apellido {
-    color: #e0e0e0;
-}
-
-/* Botones de acción */
-.acciones-buttons {
-    display: flex;
-    gap: 0.5rem;
-}
-
-.btn-accion {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 2rem;
-    height: 2rem;
-    border-radius: 0.4rem;
-    text-decoration: none;
-    transition: all 0.2s ease;
-}
-
-.btn-accion img {
-    width: 1rem;
-    height: 1rem;
-}
-
-.btn-accion.editar {
-    background-color: rgba(52, 152, 219, 0.2);
-    border: 1px solid rgba(52, 152, 219, 0.3);
-}
-
-.btn-accion.editar:hover {
-    background-color: rgba(52, 152, 219, 0.3);
-    transform: scale(1.05);
-}
-
-.btn-accion.editar img {
-    filter: brightness(0) saturate(100%) invert(64%) sepia(98%) saturate(2945%) hue-rotate(189deg) brightness(96%) contrast(87%);
-}
-
-.btn-accion.eliminar {
-    background-color: rgba(231, 76, 60, 0.2);
-    border: 1px solid rgba(231, 76, 60, 0.3);
-}
-
-.btn-accion.eliminar:hover {
-    background-color: rgba(231, 76, 60, 0.3);
-    transform: scale(1.05);
-}
-
-.btn-accion.eliminar img {
-    filter: brightness(0) saturate(100%) invert(27%) sepia(51%) saturate(2878%) hue-rotate(346deg) brightness(104%) contrast(97%);
-}
-
-/* Estado sin usuarios */
-.no-users {
-    text-align: center;
-    padding: 3rem 2rem;
-    color: #d4d4d4;
-}
-
-.no-users-icon {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 4rem;
-    height: 4rem;
-    margin: 0 auto 1.5rem;
-    background-color: rgba(255, 255, 255, 0.1);
-    border-radius: 1rem;
-}
-
-.no-users-icon img {
-    width: 2rem;
-    height: 2rem;
-    filter: brightness(0) invert(1);
-    opacity: 0.6;
-}
-
-.no-users h3 {
-    font-size: 1.3rem;
-    font-weight: 600;
-    margin-bottom: 0.5rem;
-    color: #f3f3f3;
-}
-
-.no-users p {
-    font-size: 1rem;
-    margin-bottom: 1.5rem;
-    opacity: 0.8;
-}
-
-.btn-primary {
-    display: inline-flex;
-    align-items: center;
-    gap: 0.5rem;
-    padding: 0.8rem 1.5rem;
-    background-color: #ffffff;
-    color: #000;
-    text-decoration: none;
-    border-radius: 0.6rem;
-    font-weight: 600;
-    transition: all 0.2s ease;
-}
-
-.btn-primary:hover {
-    background-color: #f0f0f0;
-    transform: translateY(-2px);
-}
-
-.btn-primary img {
-    width: 1rem;
-    height: 1rem;
-    filter: brightness(0);
-}
-
-/* Responsive */
-@media (max-width: 768px) {
-    .actions-container {
-        grid-template-columns: 1fr;
-    }
-    
-    .action-card {
-        flex-direction: column;
-        text-align: center;
-        gap: 1rem;
-    }
-    
-    .action-icon {
-        margin-right: 0;
-        margin-bottom: 0.5rem;
-    }
-    
-    .search-input-container {
-        width: 100%;
-    }
-    
-    #searchUsers {
-        width: 100%;
-    }
-    
-    .usuarios-table th,
-    .usuarios-table td {
-        padding: 0.75rem 0.5rem;
-        font-size: 0.8rem;
-    }
-    
-    .th-content {
-        flex-direction: column;
-        gap: 0.3rem;
-    }
-    
-    .th-content img {
-        width: 0.8rem;
-        height: 0.8rem;
-    }
-}
-
-@media (max-width: 500px) {
-    .usuarios-container {
-        padding: 0.5rem;
-    }
-    
-    .usuarios-table-container {
-        padding: 1rem;
-    }
-    
-    .acciones-buttons {
-        flex-direction: column;
-        gap: 0.3rem;
-    }
-    
-    .btn-accion {
-        width: 1.8rem;
-        height: 1.8rem;
-    }
-    
-    .btn-accion img {
-        width: 0.9rem;
-        height: 0.9rem;
-    }
-}
