--- frontend/app/dashboard/Admin/lotes/Users/eliminar.php
@@ -1,9 +1,0 @@
-<?php
-require_once 'UsuarioModel.php';
-$model = new UsuarioModel();
-
-$id = $_GET['id'] ?? null;
-if ($id) {
-    $model->eliminar($id);
-}
-header('Location: UserList.php');
