+++ frontend/app/dashboard/Admin/modules/GestionLotes/carnet/carnet.php
@@ -1,0 +1,2 @@
+<h1>Anasheiii</h1>
+<p>Esto va a andar flama, creanme.<br>- Panchito | 01/07/2025 - 15:26</p>
