+++ frontend/app/dashboard/Admin/modules/GestionLotes/usuarios/eliminar.php
@@ -1,0 +1,9 @@
+<?php
+require_once 'UsuarioModel.php';
+$model = new UsuarioModel();
+
+$id = $_GET['id'] ?? null;
+if ($id) {
+    $model->eliminar($id);
+}
+header('Location: UserList.php');
